Added initial jQuery Mobile support based on Kris Borcher's work.
Refined the AngularJS directive to avoid id references to page content and without requiring patching of AngularJS to broadcast events on $rootscope.

Refined the markup of the search and detail pages to use jQuery widgets. Note that ng-show requires a wrapper since jQM enhances the DOM before AngularJS decides to show or hide an element.